LPS-171321 Migrate FDS Autocomplete filters to Selection filters in commerce
