+++ modules/apps/commerce/commerce-frontend-js/dev/components/DatasetDisplay.js
@@ -332,23 +332,25 @@ const ordersDataSetProps = {
 	filters: [
 		{
 			apiURL: '/o/headless-commerce-admin-account/v1.0/accounts',
+			autocompleteEnabled: true,
 			id: 'accountId',
 			inputPlaceholder: 'Search for account',
 			itemKey: 'id',
 			itemLabel: 'name',
 			label: 'Account',
-			selectionType: 'single',
-			type: 'autocomplete',
+			multiple: false,
+			type: 'selection',
 		},
 		{
 			apiURL: '/o/headless-commerce-admin-channel/v1.0/channels',
+			autocompleteEnabled: true,
 			id: 'channelId',
 			inputPlaceholder: 'Search for Channel',
 			itemKey: 'id',
 			itemLabel: 'name',
 			label: 'Channel',
-			selectionType: 'single',
-			type: 'autocomplete',
+			multiple: false,
+			type: 'selection',
 		},
 		{
 			id: 'createDate',
