+++ modules/apps/commerce/commerce-order-web/src/main/java/com/liferay/commerce/order/web/internal/frontend/data/set/filter/CommerceAccountSelectionFDSFilter.java
@@ -12,32 +12,37 @@
  * details.
  */
 
-package com.liferay.commerce.pricing.web.internal.frontend.data.set.filter;
+package com.liferay.commerce.order.web.internal.frontend.data.set.filter;
 
-import com.liferay.commerce.pricing.web.internal.constants.CommercePricingFDSNames;
-import com.liferay.frontend.data.set.filter.BaseAutocompleteFDSFilter;
+import com.liferay.commerce.order.web.internal.constants.CommerceOrderFDSNames;
+import com.liferay.frontend.data.set.constants.FDSEntityFieldTypes;
+import com.liferay.frontend.data.set.filter.BaseSelectionFDSFilter;
 import com.liferay.frontend.data.set.filter.FDSFilter;
 
 import org.osgi.service.component.annotations.Component;
 
 /**
- * @author Alessio Antonio Rendina
+ * @author Marco Leo
  */
 @Component(
-	property = "frontend.data.set.name=" + CommercePricingFDSNames.PRICE_LISTS,
+	property = "frontend.data.set.name=" + CommerceOrderFDSNames.ALL_ORDERS,
 	service = FDSFilter.class
 )
-public class CommerceCatalogAutocompleteFDSFilter
-	extends BaseAutocompleteFDSFilter {
+public class CommerceAccountSelectionFDSFilter extends BaseSelectionFDSFilter {
 
 	@Override
 	public String getAPIURL() {
-		return "/o/headless-commerce-admin-catalog/v1.0/catalogs?sort=name:asc";
+		return "/o/headless-commerce-admin-account/v1.0/accounts?sort=name:asc";
+	}
+
+	@Override
+	public String getEntityFieldType() {
+		return FDSEntityFieldTypes.COLLECTION;
 	}
 
 	@Override
 	public String getId() {
-		return "catalogId";
+		return "accountId";
 	}
 
 	@Override
@@ -52,11 +57,16 @@ public class CommerceCatalogAutocompleteFDSFilter
 
 	@Override
 	public String getLabel() {
-		return "catalog";
+		return "account";
 	}
 
 	@Override
-	public boolean isMultipleSelection() {
+	public boolean isAutocompleteEnabled() {
+		return true;
+	}
+
+	@Override
+	public boolean isMultiple() {
 		return true;
 	}
