+++ modules/apps/commerce/commerce-order-web/src/main/java/com/liferay/commerce/order/web/internal/frontend/data/set/filter/CommerceOrderStatusSelectionFDSFilter.java
@@ -36,7 +36,8 @@ import org.osgi.service.component.annotations.Reference;
 	property = "frontend.data.set.name=" + CommerceOrderFDSNames.ALL_ORDERS,
 	service = FDSFilter.class
 )
-public class CommerceOrderStatusFDSFilter extends BaseSelectionFDSFilter {
+public class CommerceOrderStatusSelectionFDSFilter
+	extends BaseSelectionFDSFilter {
 
 	@Override
 	public String getEntityFieldType() {
