+++ modules/apps/commerce/commerce-pricing-web/src/main/java/com/liferay/commerce/pricing/web/internal/frontend/data/set/filter/CommerceChannelSelectionFDSFilter.java
@@ -12,32 +12,40 @@
  * details.
  */
 
-package com.liferay.commerce.product.definitions.web.internal.frontend.data.set.filter;
+package com.liferay.commerce.pricing.web.internal.frontend.data.set.filter;
 
-import com.liferay.commerce.product.definitions.web.internal.constants.CommerceProductFDSNames;
-import com.liferay.frontend.data.set.filter.BaseAutocompleteFDSFilter;
+import com.liferay.commerce.pricing.web.internal.constants.CommercePricingFDSNames;
+import com.liferay.frontend.data.set.constants.FDSEntityFieldTypes;
+import com.liferay.frontend.data.set.filter.BaseSelectionFDSFilter;
 import com.liferay.frontend.data.set.filter.FDSFilter;
 
 import org.osgi.service.component.annotations.Component;
 
 /**
- * @author Marco Leo
+ * @author Alessio Antonio Rendina
  */
 @Component(
-	property = "frontend.data.set.name=" + CommerceProductFDSNames.PRODUCT_DEFINITIONS,
+	property = {
+		"frontend.data.set.name=" + CommercePricingFDSNames.DISCOUNTS,
+		"frontend.data.set.name=" + CommercePricingFDSNames.PRICE_LISTS
+	},
 	service = FDSFilter.class
 )
-public class CommerceCatalogAutocompleteFDSFilter
-	extends BaseAutocompleteFDSFilter {
+public class CommerceChannelSelectionFDSFilter extends BaseSelectionFDSFilter {
 
 	@Override
 	public String getAPIURL() {
-		return "/o/headless-commerce-admin-catalog/v1.0/catalogs?sort=name:asc";
+		return "/o/headless-commerce-admin-channel/v1.0/channels?sort=name:asc";
+	}
+
+	@Override
+	public String getEntityFieldType() {
+		return FDSEntityFieldTypes.COLLECTION;
 	}
 
 	@Override
 	public String getId() {
-		return "catalogId";
+		return "channelId";
 	}
 
 	@Override
@@ -52,12 +60,17 @@ public class CommerceCatalogAutocompleteFDSFilter
 
 	@Override
 	public String getLabel() {
-		return "catalog";
+		return "channel";
 	}
 
 	@Override
-	public boolean isMultipleSelection() {
-		return false;
+	public boolean isAutocompleteEnabled() {
+		return true;
+	}
+
+	@Override
+	public boolean isMultiple() {
+		return true;
 	}
 
 }
